client/requirements.txt
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-decouple==3.6
Werkzeug==2.2.2
cmake
dlib
face-recognition
gunicorn
DateTime
PyJWT
bcrypt